requirements.txt
#
# Requirements for running drf-cached-instances
#

Django==1.7.1
pytz==2014.9
djangorestframework==2.4.4

#
# Requirements for developing drf-cached-instances
#

# Packaging
wheel==0.23.0

# PEP8, PEP257, and static analysis
mccabe==0.2.1
pep8==1.5.7
pyflakes==0.8.1
flake8==2.2.5
pep257==0.3.2
flake8-docstrings==0.2.1

# Code coverage
coverage==3.7.1

# Better test runner
nose==1.3.4
# Until django-nose==1.3 is release to PyPI
https://github.com/jwhitlock/django-nose/archive/django-1.7.zip

# Django extensions
six==1.8.0
django-extensions==1.4.5

# Documentation
Pygments==1.6
Sphinx==1.2.2
docutils==0.12

# Packaging Linters
check-manifest==0.21
pyroma==1.7

# Multi-env test runner
virtualenv==1.11.6
py==1.4.26
tox==1.8.1

# Better interactive debugging
gnureadline==6.3.3
ipython==2.3.0
ipdb==0.8
ipdbplugin==1.4.1

# Production quality serving of static files
static3==0.5.1
dj-static==0.0.6

# ./manage.py runserveR_plus for better tracebacks
Werkzeug==0.9.6

# Django debug toolbar
sqlparse==0.1.13
django-debug-toolbar==1.2.2

# Testing library, added as unittest.mock in Python 3.3
mock==1.0.1

#
# Requirements for Heroku
#

# WSGI runner
gunicorn==19.1.1

# PostgreSQL driver
psycopg2==2.5.4
